refactor(chess/application): console용 class와 web용 class 패키지 분리
