+++ src/main/java/chess/application/web/WebApplication.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.application.web;
 
 import static spark.Spark.exception;
 import static spark.Spark.get;
